requirements.txt
fastapi
uvicorn
faster-whisper
groq
elevenlabs
python-dotenv
websockets
numpy
sounddevice
python-multipart
scipy